Perf: encode ring_id in task_id and fix AICore cache flush ordering (#316)
PerfRecord.task_id was uint32 and could only represent local task
indices, making it impossible to distinguish tasks across rings in
multi-ring runtimes (tensormap_and_ringbuffer, aicpu_build_graph).

- Widen PerfRecord.task_id uint32 → uint64; encode as (ring_id << 32)
  | local_id; remove now-redundant ring_id field
- Widen PerfRecord.fanout[] int32 → uint64 to store full 64-bit
  successor task_ids across rings
- Merge AicpuPhaseRecord.tasks_processed/padding into union
  { task_id; tasks_processed } for orchestrator cross-view correlation
- AICPU executors overwrite task_id with full PTO2 encoding after
  handshake match; host_build_graph sets task_id to plain integer index
- Rename PTO2TaskDescriptor.mixed_task_id → task_id; pass task_id.raw
  to orchestrator perf phases (replacing local_id)
- Update host perf export to derive ring_id from task_id >> 32
- Add format_task_display() in swimlane_converter.py for r{ring}t{local}
  human-readable labels in Chrome trace output
- AICore perf flush: replace dcci(count) + dcci(record) with
  dcci(record) + dsb(0) to ensure record visibility before count update

+++ src/a2a3/platform/include/aicpu/performance_collector_aicpu.h
@@ -101,12 +101,14 @@ void perf_aicpu_init_phase_profiling(Runtime* runtime, int num_sched_threads, in
  * @param start_time Phase start timestamp
  * @param end_time Phase end timestamp
  * @param loop_iter Current loop iteration number
- * @param tasks_processed Number of tasks processed in this phase
+ * @param tasks_processed Number of tasks processed in this batch (scheduler phases), or
+ *                        full PTO2 task_id encoding (ring_id << 32) | local_id (orchestrator
+ *                        phases in multi-ring runtimes: tensormap_and_ringbuffer, aicpu_build_graph)
  */
 void perf_aicpu_record_phase(int thread_idx,
                               AicpuPhaseId phase_id,
                               uint64_t start_time, uint64_t end_time,
-                              uint32_t loop_iter, uint32_t tasks_processed);
+                              uint32_t loop_iter, uint64_t tasks_processed);
 
 /**
  * Write orchestrator cumulative summary
@@ -138,11 +140,13 @@ void perf_aicpu_set_orch_thread_idx(int thread_idx);
  * @param start_time Phase start timestamp
  * @param end_time Phase end timestamp
  * @param submit_idx Task submission index (acts as loop_iter)
- * @param task_id Task ID (stored in tasks_processed field for task tracking)
+ * @param task_id Task identifier. For multi-ring runtimes (tensormap_and_ringbuffer, aicpu_build_graph), this is the full PTO2 encoding:
+ *               (ring_id << 32) | local_id, enabling cross-view correlation between orchestrator
+ *               and scheduler swimlanes.
  */
 void perf_aicpu_record_orch_phase(AicpuPhaseId phase_id,
                                    uint64_t start_time, uint64_t end_time,
-                                   uint32_t submit_idx, uint32_t task_id);
+                                   uint32_t submit_idx, uint64_t task_id);
 
 /**
  * Write core-to-thread assignment mapping to shared memory
